Use new `langchain-openai` partner package (#653)
* Import ChatOpenAI and AzureChatOpenAI from the updated langchain location:

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update pyproject.toml, add `langchain-openai` as an optional dependency

* remove direct dependency on openai and tiktoken packages

* move openai providers to separate module to keep langchain_openai optional

* update openai provider dependencies in docs

* pre-commit

---------
