+++ packages/jupyter-ai/pyproject.toml
@@ -26,7 +26,6 @@ dependencies = [
     "jupyterlab~=4.0",
     "aiosqlite>=0.18",
     "importlib_metadata>=5.2.0",
-    "tiktoken",                  # required for OpenAIEmbeddings
     "jupyter_ai_magics",
     "dask[distributed]",
     "faiss-cpu",                 # Not distributed by official repo
